requirements.txt
aniso8601==10.0.0
attrs==25.1.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
pymongo==4.11.1
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
typing_extensions==4.12.2
Werkzeug==3.1.3
